improve logging, split out layouts file
